refactor: calculateLottoResult 메소드명 수정 및 ENUM.calculateLottoResult 추가
